+++ packages/client/src/Call.ts
@@ -8,6 +8,7 @@ import {
   Publisher,
   Subscriber,
   toRtcConfiguration,
+  TrackPublishOptions,
   trackTypeToParticipantStreamKey,
 } from './rtc';
 import {
@@ -874,8 +875,6 @@ export class Call {
       throw new Error(`Illegal State: call.join() shall be called only once`);
     }
 
-    this.state.setCallingState(CallingState.JOINING);
-
     // we will count the number of join failures per SFU.
     // once the number of failures reaches 2, we will piggyback on the `migrating_from`
     // field to force the coordinator to provide us another SFU
@@ -905,8 +904,6 @@ export class Call {
         }
 
         if (attempt === maxJoinRetries - 1) {
-          // restore the previous call state if the join-flow fails
-          this.state.setCallingState(callingState);
           throw err;
         }
       }
@@ -981,6 +978,7 @@ export class Call {
           })
         : previousSfuClient;
     this.sfuClient = sfuClient;
+    this.unifiedSessionId ??= sfuClient.sessionId;
     this.dynascaleManager.setSfuClient(sfuClient);
 
     const clientDetails = await getClientDetails();
@@ -1008,6 +1006,7 @@ export class Call {
       try {
         const { callState, fastReconnectDeadlineSeconds, publishOptions } =
           await sfuClient.join({
+            unifiedSessionId: this.unifiedSessionId,
             subscriberSdp,
             publisherSdp,
             clientDetails,
@@ -1061,6 +1060,7 @@ export class Call {
         statsOptions,
         publishOptions: this.currentPublishOptions || [],
         closePreviousInstances: !performingMigration,
+        unifiedSessionId: this.unifiedSessionId,
       });
     }
 
@@ -1222,6 +1222,7 @@ export class Call {
     clientDetails: ClientDetails;
     publishOptions: PublishOption[];
     closePreviousInstances: boolean;
+    unifiedSessionId: string;
   }) => {
     const {
       sfuClient,
@@ -1230,6 +1231,7 @@ export class Call {
       statsOptions,
       publishOptions,
       closePreviousInstances,
+      unifiedSessionId,
     } = opts;
     const { enable_rtc_stats: enableTracing } = statsOptions;
     if (closePreviousInstances && this.subscriber) {
@@ -1286,9 +1288,9 @@ export class Call {
     }
 
     this.tracer.setEnabled(enableTracing);
+    this.sfuStatsReporter?.flush();
     this.sfuStatsReporter?.stop();
     if (statsOptions?.reporting_interval_ms > 0) {
-      this.unifiedSessionId ??= sfuClient.sessionId;
       this.sfuStatsReporter = new SfuStatsReporter(sfuClient, {
         clientDetails,
         options: statsOptions,
@@ -1298,7 +1300,7 @@ export class Call {
         camera: this.camera,
         state: this.state,
         tracer: this.tracer,
-        unifiedSessionId: this.unifiedSessionId,
+        unifiedSessionId,
       });
       this.sfuStatsReporter.start();
     }
@@ -1770,9 +1772,14 @@ export class Call {
    *
    * @param mediaStream the media stream to publish.
    * @param trackType the type of the track to announce.
+   * @param options the publish options.
    */
-  publish = async (mediaStream: MediaStream, trackType: TrackType) => {
-    if (!this.sfuClient) throw new Error(`Call not joined yet.`);
+  publish = async (
+    mediaStream: MediaStream,
+    trackType: TrackType,
+    options?: TrackPublishOptions,
+  ) => {
+    if (!this.sfuClient) throw new Error(`Call is not joined yet`);
     // joining is in progress, and we should wait until the client is ready
     await this.sfuClient.joinTask;
 
@@ -1797,15 +1804,16 @@ export class Call {
     }
 
     pushToIfMissing(this.trackPublishOrder, trackType);
-    await this.publisher.publish(track, trackType);
+    await this.publisher.publish(track, trackType, options);
 
     const trackTypes = [trackType];
     if (trackType === TrackType.SCREEN_SHARE) {
       const [audioTrack] = mediaStream.getAudioTracks();
       if (audioTrack) {
-        pushToIfMissing(this.trackPublishOrder, TrackType.SCREEN_SHARE_AUDIO);
-        await this.publisher.publish(audioTrack, TrackType.SCREEN_SHARE_AUDIO);
-        trackTypes.push(TrackType.SCREEN_SHARE_AUDIO);
+        const screenShareAudio = TrackType.SCREEN_SHARE_AUDIO;
+        pushToIfMissing(this.trackPublishOrder, screenShareAudio);
+        await this.publisher.publish(audioTrack, screenShareAudio, options);
+        trackTypes.push(screenShareAudio);
       }
     }