+++ packages/client/src/devices/ScreenShareState.ts
@@ -1,9 +1,10 @@
 import { BehaviorSubject, distinctUntilChanged } from 'rxjs';
-import { InputMediaDeviceManagerState } from './InputMediaDeviceManagerState';
+import { AudioDeviceManagerState } from './AudioDeviceManagerState';
+import { AudioBitrateProfile } from '../gen/video/sfu/models/models';
 import type { ScreenShareSettings } from '../types';
 import { RxUtils } from '../store';
 
-export class ScreenShareState extends InputMediaDeviceManagerState<DisplayMediaStreamOptions> {
+export class ScreenShareState extends AudioDeviceManagerState<DisplayMediaStreamOptions> {
   private audioEnabledSubject = new BehaviorSubject<boolean>(true);
   private settingsSubject = new BehaviorSubject<
     ScreenShareSettings | undefined
@@ -22,9 +23,16 @@ export class ScreenShareState extends InputMediaDeviceManagerState<DisplayMediaS
   settings$ = this.settingsSubject.asObservable();
 
   /**
+   * Constructs a new ScreenShareState instance.
+   */
+  constructor() {
+    super('stop-tracks', undefined, AudioBitrateProfile.MUSIC_HIGH_QUALITY);
+  }
+
+  /**
    * @internal
    */
-  protected getDeviceIdFromStream = (
+  protected override getDeviceIdFromStream = (
     stream: MediaStream,
   ): string | undefined => {
     const [track] = stream.getTracks();