+++ packages/client/src/rtc/Subscriber.ts
@@ -1,7 +1,5 @@
-import {
-  BasePeerConnection,
-  type BasePeerConnectionOpts,
-} from './BasePeerConnection';
+import { BasePeerConnection } from './BasePeerConnection';
+import type { BasePeerConnectionOpts } from './types';
 import { NegotiationError } from './NegotiationError';
 import { PeerType } from '../gen/video/sfu/models/models';
 import { SubscriberOffer } from '../gen/video/sfu/event/events';