+++ packages/react-native-sdk/ios/StreamVideoReactNative.h
@@ -9,4 +9,8 @@
 
 + (void)setup DEPRECATED_MSG_ATTRIBUTE("No need to use setup() anymore");
 
++ (BOOL)shouldRejectCallWhenBusy;
+
++ (BOOL)hasAnyActiveCall;
+
 @end