+++ sample-apps/react-native/dogfood/src/screens/AudioRoom/Room.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import {
   CallingState,
+  callManager,
   OwnCapability,
   SfuModels,
   useCall,
@@ -8,7 +9,6 @@ import {
 } from '@stream-io/video-react-native-sdk';
 import { StyleSheet } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import InCallManager from 'react-native-incall-manager';
 import { ControlsPanel } from '../../components/AudioRoom/ControlsPanel';
 import { PermissionRequestsPanel } from '../../components/AudioRoom/PermissionRequestsPanel';
 import { ParticipantsPanel } from '../../components/AudioRoom/ParticipantsPanel';
@@ -19,10 +19,9 @@ export default function Room({ onClose }: { onClose: () => void }) {
   const callingState = useCallCallingState();
   const call = useCall();
 
-  // Automatically route audio to ear piece.
   useEffect(() => {
-    InCallManager.start({ media: 'audio' });
-    return () => InCallManager.stop();
+    callManager.start();
+    return () => callManager.stop();
   }, []);
 
   // when the component unmounts, leave the call